Move stryker4s.extension.exception to stryker4s.exception
