+++ modules/core/src/main/scala/stryker4s/exception/stryker4sException.scala
@@ -1,4 +1,4 @@
-package stryker4s.extension.exception
+package stryker4s.exception
 
 import cats.data.NonEmptyList
 import cats.syntax.foldable.*
