+++ modules/core/src/main/scala/stryker4s/run/MutantRunner.scala
@@ -8,8 +8,8 @@ import fs2.io.file.{Files, Path}
 import fs2.{text, Pipe, Stream}
 import mutationtesting.{MutantResult, MutantStatus}
 import stryker4s.config.Config
+import stryker4s.exception.{InitialTestRunFailedException, UnableToFixCompilerErrorsException}
 import stryker4s.extension.FileExtensions.*
-import stryker4s.extension.exception.{InitialTestRunFailedException, UnableToFixCompilerErrorsException}
 import stryker4s.files.FilesFileResolver
 import stryker4s.log.Logger
 import stryker4s.model.*
