+++ modules/core/src/test/scala/stryker4s/mutants/tree/MutantInstrumenterTest.scala
@@ -2,8 +2,8 @@ package stryker4s.mutants.tree
 
 import cats.data.{NonEmptyList, NonEmptyVector}
 import fs2.io.file.Path
+import stryker4s.exception.UnableToBuildPatternMatchException
 import stryker4s.extension.TreeExtensions.*
-import stryker4s.extension.exception.UnableToBuildPatternMatchException
 import stryker4s.model.*
 import stryker4s.mutants.applymutants.ActiveMutationContext
 import stryker4s.mutation.{ConditionalTrue, GreaterThan, Mutation, True}
