+++ modules/sbt/src/main/scala/stryker4s/files/SbtMutatesResolver.scala
@@ -5,7 +5,7 @@ import fs2.Stream
 import fs2.io.file.Path
 import sbt.Keys.sources
 import sbt.{Compile, Project, State, Value}
-import stryker4s.extension.exception.TestSetupException
+import stryker4s.exception.TestSetupException
 
 class SbtMutatesResolver(state: State, target: Path) extends MutatesFileResolver {
